Convert reducer code to TypeScript
- Convert reducer code to TypeScript

- Move utils to separate file
