--- src/store/context.js
@@ -1,14 +1,0 @@
-import React, { useReducer } from 'react';
-
-import { defaultState, reducer } from './reducer';
-
-export const StoreContext = React.createContext();
-
-export const StoreProvider = ({ children }) => {
-  const [state, dispatch] = useReducer(reducer, defaultState);
-  return (
-    <StoreContext.Provider value={{ ...state, dispatch }}>
-      {children}
-    </StoreContext.Provider>
-  );
-};
